Added postinstall_script in pkginfo
There is an embedded tool at /Applications/VMWare Fusion.app/Contents/Library/Initialize VMWare Fusion.tool.
There is a note in the tool how it can be used.  The second case is ideal for deployment via munki, especially in non-admin environments:
```
# 2) Unattended, by the system administrator of an organization who mass
#    deploys the application.
#
# After the application has been copied to a new machine, the system
# administrator invokes this executable as root with the first argument set to
# "set". After that, any user of the organization who starts the application on
# this machine will not be prompted for root credentials.
```

I've included this postinstall_script in my autopkg override of the recipe but wanted to pass it on in case you wanted to add it to the parent recipe.  I'm not sure if there'd be a use case where the admin prompt would be desired.

The output of the .tool when run is:

```    
*** Creating directory /Library/Application Support/VMware...
    *** Setting permissions on /Library/Application Support/VMware...
    *** Creating directory /Library/Application Support/VMware/VMware Fusion...
    *** Setting permissions on /Library/Application Support/VMware/VMware Fusion...
    *** Creating directory /Library/Application Support/VMware/VMware Fusion/AdminWritable...
    *** Setting permissions on /Library/Application Support/VMware/VMware Fusion/AdminWritable...
    *** Creating directory /Library/Application Support/VMware/VMware Fusion/Shared...
    *** Setting permissions on /Library/Application Support/VMware/VMware Fusion/Shared...
    *** Setting file permissions...
```